ml-service/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
scikit-learn>=1.3.0
tensorflow>=2.15.0
python-dotenv>=1.0.0
